api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.3.0
onnxruntime==1.18.0
transformers==4.41.2
numpy==1.26.4
redis[hiredis]==5.0.6
azure-servicebus==7.12.2
azure-monitor-opentelemetry==1.6.1
opentelemetry-instrumentation-fastapi~=0.47b0
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==23.2.1
python-multipart==0.0.9
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0